Add files via upload
Found a couple sections of the code that caused some issues. The first is in regards to the path location of the rushOrderPrices.txt, the path to the file wasn't quite correct and it couldn't find it, so the rush cost was never changed from 0 thus not adding the rush cost in the final total. I changed the filepath to use a more relative approach making it PC agnostic and it now finds the file fine.

The second issue I found was actually with the original main branch code where the desk cost is being calculated. As part of the cost its should be $1 per sq. in. over 1000 sq. in. I added in the code needed to resolve this issue and everything seems to compute fine now.